Change sales and price columns to FLOAT type

+++ Scripts/ddl_bronze.sql
@@ -50,9 +50,9 @@ CREATE TABLE bronze.crm_sales_details (
     sls_order_dt INT,
     sls_ship_dt  INT,
     sls_due_dt   INT,
-    sls_sales    INT,
+    sls_sales    FLOAT,
     sls_quantity INT,
-    sls_price    INT,
+    sls_price    FLOAT,
 	Order_year	 INT
 );
 GO
